:truck: Move `attach-target-profile` to `certification/configuration`
Déplacement des fichiers attach-target-profile de complementary-certification vers configuration pour une meilleure organisation du code.

Fichiers déplacés :
- attach-target-profile-controller.js et ses tests
- attach-target-profile-route.js et ses tests
- get-complementary-certification-for-target-profile-attachment.js
- complementary-certification-badge-serializer.js et ses tests

Mise à jour des routes dans server.js pour refléter le nouveau contexte.
