+++ api/server.js
@@ -6,11 +6,9 @@ import { setupErrorHandling } from './config/server-setup-error-handling.js';
 import { databaseConnections } from './db/database-connections.js';
 import { knex } from './db/knex-database-connection.js';
 import { bannerRoutes } from './src/banner/routes.js';
+import { complementaryCertificationRoutes } from './src/certification/complementary-certification/routes.js';
 import {
   attachTargetProfileRoutes,
-  complementaryCertificationRoutes,
-} from './src/certification/complementary-certification/routes.js';
-import {
   certificationConfigurationRoutes,
   flashAssessmentConfigurationRoutes,
   scoWhitelistRoutes,
