+++ api/src/certification/complementary-certification/routes.js
@@ -1,7 +1,5 @@
-import * as attachTargetProfile from './application/attach-target-profile-route.js';
 import * as complementaryCertification from './application/complementary-certification-route.js';
 
-const attachTargetProfileRoutes = [attachTargetProfile];
 const complementaryCertificationRoutes = [complementaryCertification];
 
-export { attachTargetProfileRoutes, complementaryCertificationRoutes };
+export { complementaryCertificationRoutes };
