+++ api/src/certification/configuration/infrastructure/serializers/complementary-certification-badge-serializer.js
@@ -1,7 +1,7 @@
 import jsonapiSerializer from 'jsonapi-serializer';
 
-import { CONCURRENCY_HEAVY_OPERATIONS } from '../../../../../shared/infrastructure/constants.js';
-import { PromiseUtils } from '../../../../../shared/infrastructure/utils/promise-utils.js';
+import { CONCURRENCY_HEAVY_OPERATIONS } from '../../../../shared/infrastructure/constants.js';
+import { PromiseUtils } from '../../../../shared/infrastructure/utils/promise-utils.js';
 
 const { Deserializer } = jsonapiSerializer;
