+++ api/tests/certification/configuration/unit/application/attach-target-profile-route_test.js
@@ -1,5 +1,5 @@
-import { attachTargetProfileController } from '../../../../../src/certification/complementary-certification/application/attach-target-profile-controller.js';
-import * as moduleUnderTest from '../../../../../src/certification/complementary-certification/application/attach-target-profile-route.js';
+import { attachTargetProfileController } from '../../../../../src/certification/configuration/application/attach-target-profile-controller.js';
+import * as moduleUnderTest from '../../../../../src/certification/configuration/application/attach-target-profile-route.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
